+++ firmware/master/stm32/dac.h
@@ -29,24 +29,30 @@ public:
     };
 
     enum Trigger {
-        Timer6,
-        Timer3,
-        Timer7,
-        Timer5,
-        Timer2,
-        Timer4,
+        TIM6,
+        TIM3,
+        TIM7,
+        TIM5,
+        TIM2,
+        TIM4,
         Exti9,
         SwTrig,
         TrigNone
     };
 
     static void init();
-    static void configureChannel(int ch, Waveform waveform = WaveNone, uint8_t mask_amp = 0, Trigger trig = TrigNone, BufferMode buffmode = BufferEnabled);
+    static void configureChannel(int ch, Trigger trig = TrigNone,
+        Waveform waveform = WaveNone, uint8_t mask_amp = 0, BufferMode buffmode = BufferEnabled);
+
     static void enableChannel(int ch);
     static void disableChannel(int ch);
 
+    static void enableDMA(int ch);
+    static void disableDMA(int ch);
+
     static void writeDual(uint16_t data);
     static void write(int ch, uint16_t data, DataFormat format = RightAlign12Bit);
+    static uintptr_t address(int ch, DataFormat format = RightAlign12Bit);
 };
 
 #endif // DAC_H_